Move face analyzer to own file
